+++ mysite/templates/theme/import_mapping.html
@@ -4,9 +4,6 @@
 		<input type="hidden" name="import" value="{{import}}">
 		<div class="box-header with-border">
 			<h3 class="box-title">Daten importieren</h3>
-			<div class="box-tools pull-right">
-				<button class="btn btn-box-tool" data-widget="collapse"><span class="fa fa-plus"></span></button>
-			</div>
 		</div>
 		<div class="box-body">
 			<div class="table-responsive">
