+++ mysite/templates/theme/konfiguration.html
@@ -100,40 +100,3 @@
 		</form>
 	</div>
 </div>
-
-  <div class="theme-double-columns">
-	<div class="box">
-		<div class="box-header">
-			<h3 class="box-title">Kategorien installieren</h3>
-		</div>
-		<form role="form" action="/configuration/" method="post">{% csrf_token %}
-			<div class="box-body">
-				<div class="form-group">
-					<label>Server</label> <input type="text" name="server" class="form-control"/>
-					<label>Email</label> <input type="text" name="email" class="form-control"/>
-					<label>Passwort</label> <input type="password" name="password" class="form-control"/>
-				</div>
-				</div>
-			<div class="box-footer">
-				<button type="submit" name="action" value="set_kategorien" id="set_kategorien" class="btn bordered-hover">Daten aktualisieren</button>
-			</div>
-		</form>
-	</div>
-	<div class="box">
-		<div class="box-header">
-			<h3 class="box-title">Daten importieren</h3>
-		</div>
-		<form role="form" action="/configuration/" method="post">{% csrf_token %}
-			<div class="box-body">
-				<div class="form-group">
-					<label>Server</label> <input type="text" name="server" class="form-control"/>
-					<label>Email</label> <input type="text" name="email" class="form-control"/>
-					<label>Passwort</label> <input type="password" name="password" class="form-control"/>
-				</div>
-				</div>
-			<div class="box-footer">
-				<button type="submit" name="action" value="load_online_transactions" id="load_online_transactions" class="btn bordered-hover">Daten importieren</button>
-			</div>
-		</form>
-	</div>
-</div>
